Apply requested changes from gabber235 for MythicMobsAdapter
I read some of the comments of gabber235 for https://github.com/gabber235/TypeWriter/pull/41 and decided to attempt to apply them. Forgive me if there are some mistakes, as I am not that experienced with Java/Kotlin.

Here's what I changed:
- All comments related to the MythicMobsAdapter
- Comments related to minor changes in the RPGRegionsAdapter, such as typos, extra whitespaces, or renaming
- All comments related to documentation

I wasn't able to make the major changes with the RPGRegionsAdapter, as I am not too familiar with this plugin and did not have enough time.

Peace!